requirements.txt
# Core Django
Django>=4.2,<5.0

# Django REST Framework
djangorestframework>=3.14.0

# # Optional: For JWT support (currently not used but useful if enabled later)
# djangorestframework-simplejwt>=5.2.2

# # For caching with Redis (optional, used if cache backend is Redis)
# django-redis>=5.3.0

# # Dev & tooling
# black>=24.1.1
# flake8>=6.1.0
# pre-commit>=3.5.0

# # Testing
# coverage>=7.4.0

# If you’re using psycopg2 for PostgreSQL (optional)
# psycopg2-binary>=2.9.7

# If using SQLite (default), no extra DB dependencies needed
